Edit and build statistic update

+++ src/main/kotlin/com/github/dmitrypogrebnoy/feedbacktest/state/editor/EditStateConstants.kt
@@ -1,3 +1,3 @@
 package com.github.dmitrypogrebnoy.feedbacktest.state.editor
 
-internal const val DEFAULT_NUMBER_EDIT_KOTLIN_FILE: Int = 0
+internal const val DEFAULT_NUMBER_EDIT_KOTLIN_FILE: Long = 0
